Fix issue where wp-sec fails due to mismatched plugin or theme name case.
This was originally addressed in
https://github.com/markri/wp-sec/issues/14, but the problem with that
approach is the lowercase flag is all-or-nothing. That means that it
will fix the issue with Avada, but it will then break for plugins or
themes where the name is correctly uppercase in both Wordpress and
wpvulndb, like Divi.

In order to ensure we always find a result, we have to do a comparison
of the lowercase version of the plugin name to the lowercase title of
the entry from wpvulndb, which we can safely always due. The only
downside to this approach is that we have to loop through the entire set
of of results from wpvulndb rather than using array_key_exists, but the
performance hit should be very minimal considering the likely total size
of that array.